docs: rewrite comments as plain NatSpec; drop unused tinlake dep; add CI

+++ src/math/interest.sol
@@ -10,11 +10,15 @@ contract Interest is Math {
     // @param lastUpdated When the interest rate was last updated
     // @param pie Total sum of all amounts accumulating under one interest rate, divided by that rate
     // @return The new accumulated rate, as well as the difference between the debt calculated with the old and new accumulated rates.
-    function compounding(uint chi, uint ratePerSecond, uint lastUpdated, uint pie) public view returns (uint, uint) {
+    function compounding(uint256 chi, uint256 ratePerSecond, uint256 lastUpdated, uint256 pie)
+        public
+        view
+        returns (uint256, uint256)
+    {
         require(block.timestamp >= lastUpdated, "tinlake-math/invalid-timestamp");
         require(chi != 0);
         // instead of a interestBearingAmount we use a accumulated interest rate index (chi)
-        uint updatedChi = _chargeInterest(chi ,ratePerSecond, lastUpdated, block.timestamp);
+        uint256 updatedChi = _chargeInterest(chi, ratePerSecond, lastUpdated, block.timestamp);
         return (updatedChi, safeSub(rmul(updatedChi, pie), rmul(chi, pie)));
     }
 
@@ -23,25 +27,32 @@ contract Interest is Math {
     // @param ratePerSecond Interest rate accumulation per second in RAD(10ˆ27)
     // @param lastUpdated last time the interest has been charged
     // @return interestBearingAmount + interest
-    function chargeInterest(uint interestBearingAmount, uint ratePerSecond, uint lastUpdated) public view returns (uint) {
+    function chargeInterest(uint256 interestBearingAmount, uint256 ratePerSecond, uint256 lastUpdated)
+        public
+        view
+        returns (uint256)
+    {
         if (block.timestamp >= lastUpdated) {
             interestBearingAmount = _chargeInterest(interestBearingAmount, ratePerSecond, lastUpdated, block.timestamp);
         }
         return interestBearingAmount;
     }
 
-    function _chargeInterest(uint interestBearingAmount, uint ratePerSecond, uint lastUpdated, uint current) internal pure returns (uint) {
+    function _chargeInterest(uint256 interestBearingAmount, uint256 ratePerSecond, uint256 lastUpdated, uint256 current)
+        internal
+        pure
+        returns (uint256)
+    {
         return rmul(rpow(ratePerSecond, current - lastUpdated, ONE), interestBearingAmount);
     }
 
-
     // convert pie to debt/savings amount
-    function toAmount(uint chi, uint pie) public pure returns (uint) {
+    function toAmount(uint256 chi, uint256 pie) public pure returns (uint256) {
         return rmul(pie, chi);
     }
 
     // convert debt/savings amount to pie
-    function toPie(uint chi, uint amount) public pure returns (uint) {
+    function toPie(uint256 chi, uint256 amount) public pure returns (uint256) {
         return rdivup(amount, chi);
     }
